requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
pillow==5.3.0; python_version != '3.1.*'
pytelegrambotapi==3.6.6
requests==2.19.1
six==1.11.0
urllib3==1.23; python_version >= '2.6'
